Redo restacking/rebasing logic
Rather than requiring that a restack completes or rolls back in its
entirety, allow it to be abandoned at any point, keeping any rebases
that succeeded. This now requires storing a SHA along with the
branch/parent relationship so that we are able to identify the fork
point after we've changed the parent. Some additional bookkeeping will
need to be done to identify out of band rebases to a stack and update
the parent SHAs.