Stop automatically building dependencies
This is such a heavy task to run with high chance of failure (like
working offline), leaving you in a worse state than before you ran it.
With the gradle caching logic being completely unreliable at times,
this results in frequent rebuilds of these dependencies.

Move this task back to being explicit and restore the check task.